+++ src/keyboard_macro.js
@@ -100,6 +100,10 @@ const KeyboardMacro = function({ awaitController }) {
 
     const push = function(spec) {
         if (recording) {
+            if (spec.record === 'side-effect') {
+                // side-effect mode
+                return;
+            }
             sequence.push(spec);
         }
     };
@@ -283,7 +287,8 @@ const KeyboardMacro = function({ awaitController }) {
                 await playbackImpl(spec.args);
                 return;
             }
-            if (onBeginWrappedCommandCallback) {
+            const sideEffectMode = spec.record === 'side-effect';
+            if (!sideEffectMode && onBeginWrappedCommandCallback) {
                 onBeginWrappedCommandCallback();
             }
             try {
@@ -292,7 +297,7 @@ const KeyboardMacro = function({ awaitController }) {
                     push(spec);
                 }
             } finally {
-                if (onEndWrappedCommandCallback) {
+                if (!sideEffectMode && onEndWrappedCommandCallback) {
                     onEndWrappedCommandCallback();
                 }
             }